requirements.txt
confluent-kafka==2.1.1
apache-airflow-providers-apache-kafka==1.0.0
apache-airflow==2.9.1